+++ elementary_homework_2/ArithmeticMean_1.java
@@ -2,9 +2,9 @@ package ua.hillel.tatiana.elementary_homework_2;
 
 public class ArithmeticMean_1 {
     public static void main(String[] args) {
-        int sum = 0;
+        double sum = 0;
         for (String str : args) {
-            int i = Integer.valueOf(str);
+            double i = Double.valueOf(str);
             sum += i;
         }
         System.out.println("Result:" + sum / args.length);
